Restructured file tree for better navigation
