+++ src/app/app-routing.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import {LogInComponent} from "./auth/log-in/log-in.component";
-import {SchoolsListComponent} from "./schools-list/schools-list.component";
-import {MainViewComponent} from "./main-view/main-view.component";
-import {WorkordersListComponent} from "./workorders-list/workorders-list.component";
-import {AllWorkordersComponent} from "./all-workorders/all-workorders.component";
-import {OneSchoolComponent} from "./one-school/one-school.component";
-import {OneWorkorderComponent} from "./one-workorder/one-workorder.component";
-import {CreateWorkorderComponent} from "./create-workorder/create-workorder.component";
+import {SchoolsListComponent} from "./school/schools-list/schools-list.component";
+import {MainViewComponent} from "./school/main-view/main-view.component";
+import {WorkordersListComponent} from "./workorder/workorders-list/workorders-list.component";
+import {AllWorkordersComponent} from "./workorder/all-workorders/all-workorders.component";
+import {OneSchoolComponent} from "./school/one-school/one-school.component";
+import {OneWorkorderComponent} from "./workorder/one-workorder/one-workorder.component";
+import {CreateWorkorderComponent} from "./workorder/create-workorder/create-workorder.component";
 
 const routes: Routes = [
   {path: 'login', component: LogInComponent},
