+++ src/app/school/main-view/main-view.component.ts
