+++ src/app/school/schools-list/schools-list.component.css
