+++ src/app/school/schools-list/schools-list.component.ts
