+++ src/app/school/shool-info/school-info.component.css
