+++ src/app/school/shool-info/school-info.component.spec.ts
